+++ src/ui/client/mod.ts
@@ -1,3 +1,4 @@
+export { type ActiveDescendantOptions, mountActiveDescendant, resetActiveDescendant } from "./active-descendant";
 export { applyControlValue, bindField, bindGroup, parseControlValue } from "./bind";
 export { ACTIVE_COMPOSITE_ITEM, mountRovingFocus, type RovingFocusOptions } from "./composite";
 export { activeElement, asElement, closestAcross, contains, eventTarget, ownerDocument, ownerWindow } from "./dom";